backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic[email]==2.10.3
python-dotenv==1.0.1
